# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Notted API",
    "version": "1.0.0",
    "description": "The versioned public integration surface of Notted. Every path below is served under the `/api/v1` prefix and is covered by the compatibility promise of that version. tRPC is the first-party transport for the Notted web client only: it is an internal implementation detail, is not part of this public contract, and may change without a version bump. Workspace-scoped routes accept an API key bearer token; the session-only routes declare `security: []` and are reachable with a browser session alone."
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/webhooks": {
      "get": {
        "summary": "List the workspace's webhook endpoints.",
        "tags": [
          "Webhooks"
        ],
        "description": "Never returns the signing secret; only its ciphertext is stored.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                {
                  "type": "string",
                  "pattern": "^(0|[1-9]\\d*)$"
                }
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                {
                  "type": "string",
                  "pattern": "^(0|[1-9]\\d*)$"
                }
              ]
            }
          }
        ],
        "responses": {
          "2XX": {
            "description": "List the workspace's webhook endpoints.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEndpointPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Register a webhook endpoint.",
        "tags": [
          "Webhooks"
        ],
        "description": "One of only two responses that carry the raw signing secret, and it is shown exactly once. The endpoint arrives disabled and unverified.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          }
        },
        "responses": {
          "2XX": {
            "description": "Register a webhook endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookCreateResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/webhooks/{webhookId}": {
      "delete": {
        "summary": "Delete a webhook endpoint.",
        "tags": [
          "Webhooks"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "2XX": {
            "description": "Delete a webhook endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookDeleteResult"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a webhook endpoint.",
        "tags": [
          "Webhooks"
        ],
        "description": "An endpoint cannot be enabled until it is verified. Changing the URL resets both flags.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookUpdate"
              }
            }
          }
        },
        "responses": {
          "2XX": {
            "description": "Update a webhook endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEndpoint"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/webhooks/{webhookId}/rotate-secret": {
      "post": {
        "summary": "Rotate an endpoint's signing secret.",
        "tags": [
          "Webhooks"
        ],
        "description": "The other response that carries the raw signing secret, shown exactly once. The new secret signs in-flight retries immediately.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "2XX": {
            "description": "Rotate an endpoint's signing secret.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSecretRotationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/webhooks/{webhookId}/verify": {
      "post": {
        "summary": "Send the verification challenge to an endpoint.",
        "tags": [
          "Webhooks"
        ],
        "description": "Sends one signed request to the endpoint synchronously and records a delivery attempt either way. An endpoint cannot be enabled until it is verified.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "2XX": {
            "description": "Send the verification challenge to an endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookVerificationResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ntd_pk_*",
        "description": "Workspace API key, sent as `Authorization: Bearer ntd_pk_...`. A key is bound to one workspace and carries `read`, `write`, or `admin` scopes; a request is authorized only when the key's workspace matches the route's `workspaceId` and its scopes cover the action. The secret is returned once, when the key is created."
      }
    },
    "schemas": {
...
      "IsoTimestamp": {
        "type": "string",
        "format": "date-time",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
      },
...
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
      },
      "WebhookCreate": {
        "type": "object",
        "properties": {
          "url": {
            "$ref": "#/components/schemas/WebhookUrl"
          },
          "events": {
            "$ref": "#/components/schemas/WebhookEvents"
          }
        },
        "required": [
          "url",
          "events"
        ],
        "additionalProperties": false
      },
      "WebhookCreateResult": {
        "type": "object",
        "properties": {
          "webhook": {
            "$ref": "#/components/schemas/WebhookEndpoint"
          },
          "secret": {
            "$ref": "#/components/schemas/WebhookSecret"
          }
        },
        "required": [
          "webhook",
          "secret"
        ],
        "additionalProperties": false
      },
      "WebhookDeleteResult": {
        "type": "object",
        "properties": {
          "webhookId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "deleted": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "webhookId",
          "deleted"
        ],
        "additionalProperties": false
      },
      "WebhookDelivery": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "webhookId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "eventId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "event": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "status": {
            "$ref": "#/components/schemas/WebhookDeliveryStatus"
          },
          "attempt": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "responseStatus": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 100,
                "maximum": 599
              },
              {
                "type": "null"
              }
            ]
          },
          "responseBodySnippet": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          },
          "errorMessage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WebhookDeliveryErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "payloadHash": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 64,
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ]
          },
          "deliveredAt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IsoTimestamp"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          }
        },
        "required": [
          "id",
          "webhookId",
          "eventId",
          "event",
          "status",
          "attempt",
          "responseStatus",
          "responseBodySnippet",
          "errorMessage",
          "payloadHash",
          "deliveredAt",
          "createdAt"
        ],
        "additionalProperties": false
      },
      "WebhookDeliveryErrorCode": {
        "type": "string",
        "enum": [
          "timeout",
          "connection_failed",
          "dns_blocked",
          "url_rejected",
          "tls_failed",
          "http_error",
          "response_too_large",
          "resource_unavailable",
          "resource_forbidden",
          "secret_unavailable"
        ]
      },
...
      "WebhookDeliveryStatus": {
        "type": "string",
        "enum": [
          "pending",
          "success",
          "failed",
          "retrying"
        ]
      },
      "WebhookEndpoint": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "workspaceId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "url": {
            "$ref": "#/components/schemas/WebhookUrl"
          },
          "events": {
            "$ref": "#/components/schemas/WebhookEvents"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "isVerified": {
            "type": "boolean"
          },
          "createdById": {
            "$ref": "#/components/schemas/Uuid"
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "url",
          "events",
          "isEnabled",
          "isVerified",
          "createdById",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "WebhookEndpointPage": {
        "type": "object",
        "properties": {
          "items": {
            "readOnly": true,
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookEndpoint"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "page",
          "limit",
          "hasMore"
        ],
        "additionalProperties": false
      },
      "WebhookEvent": {
        "type": "string",
        "enum": [
          "note.created",
          "note.updated",
          "note.deleted",
          "project.created",
          "member.joined"
        ]
      },
      "WebhookEvents": {
        "minItems": 1,
        "maxItems": 5,
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/WebhookEvent"
        }
      },
...
      "WebhookSecret": {
        "type": "string",
        "pattern": "^whsec_[A-Za-z0-9_-]{43}$"
      },
      "WebhookSecretRotationResult": {
        "type": "object",
        "properties": {
          "webhook": {
            "$ref": "#/components/schemas/WebhookEndpoint"
          },
          "secret": {
            "$ref": "#/components/schemas/WebhookSecret"
          }
        },
        "required": [
          "webhook",
          "secret"
        ],
        "additionalProperties": false
      },
      "WebhookUpdate": {
        "type": "object",
        "properties": {
          "url": {
            "$ref": "#/components/schemas/WebhookUrl"
          },
          "events": {
            "$ref": "#/components/schemas/WebhookEvents"
          },
          "isEnabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "WebhookUrl": {
        "type": "string",
        "minLength": 1,
        "maxLength": 2048
      },
      "WebhookVerificationResult": {
        "type": "object",
        "properties": {
          "webhook": {
            "$ref": "#/components/schemas/WebhookEndpoint"
          },
          "isVerified": {
            "type": "boolean"
          },
          "delivery": {
            "$ref": "#/components/schemas/WebhookDelivery"
          }
        },
        "required": [
          "webhook",
          "isVerified",
          "delivery"
        ],
        "additionalProperties": false
      },
...
    }
  }
}